private.py
import os
from dotenv import load_dotenv

load_dotenv()

TWITTER_KEY = os.getenv('TWITTER_KEY')
TWITTER_SECRET = os.getenv('TWITTER_SECRET')
TWITTER_APP_KEY = os.getenv('TWITTER_APP_KEY')
TWITTER_APP_SECRET = os.getenv('TWITTER_APP_SECRET')
CONNECTION_STRING = os.getenv('CONNECTION_STRING')
